deleting network/p2p/ping_test.go since it introduces a depedency on go-libp2p-swarm which require go 1.14+ to compile; updating libp2pNode_test.TestPing to additionally check for version and height reported in a Ping response
